First batch of changes from Ensi
- Use double precision on functions that would only have double variant
in C (abs etc)
- Fixed raven's baddie dealing with silent printing of * in Com_Printf
and down the wire.
- Mousewheel added to console
- Shift no longer needed to open console
- Removed anti tamper code.
- Removed cd check code.
- Fixed SCR_UpdateScreen triggering SCR_DrawScreenField with non 4:3
aspect ratios thus trying to RE_SetColor(black) before the renderer was
valid in the moment of UI being restarted.
- Command Buffer size increased to 128*1024 from 16384.
- wait < 0 properly ignored.
- Command line parsing improved from ioq3 patches.
- execq added.
- echo cmd preserves colors now from ioq3.
- Upped max cvars to 2048 from 1224.
- Some misc cvar code fixes (not all yet from ioq3)
- Fixed RE_SetColor not checking for tre.registered.
- Merged in ioq3's R_PrintLongString for GL_EXTENSIONS string print that
caused a stack crash otherwise (most likely due to how many extensions
our cards have now).
- Disable the startup basic stretchpic thats not needed in q3 so why
here still ? :| (commented in q3 too)
- Added FixRenderCommandList from q3/et (Arnout) see comment for what it
does.

+++ codemp/client/cl_cin.cpp
@@ -1194,7 +1194,7 @@ e_status CIN_RunCinematic (int handle)
 	}
 
 	thisTime = Sys_Milliseconds()*com_timescale->value;
-	if (cinTable[currentHandle].shader && (abs(thisTime - (int)cinTable[currentHandle].lastTime))>100) {
+	if (cinTable[currentHandle].shader && (abs(thisTime - (double)cinTable[currentHandle].lastTime))>100) {
 		cinTable[currentHandle].startTime += thisTime - cinTable[currentHandle].lastTime;
 	}
 	cinTable[currentHandle].tfps = ((((Sys_Milliseconds()*com_timescale->value) - cinTable[currentHandle].startTime)*cinTable[currentHandle].roqFPS)/1000);
